reduce more duplication and revert changes to extHostDebugService
