+++ src/vs/platform/externalTerminal/electron-main/externalTerminalService.test.ts
@@ -5,7 +5,7 @@
 
 import { deepEqual, equal } from 'assert';
 import { DEFAULT_TERMINAL_OSX } from 'vs/platform/externalTerminal/common/externalTerminal';
-import { WindowsExternalTerminalService, MacExternalTerminalService, LinuxExternalTerminalService } from 'vs/platform/externalTerminal/electron-main/externalTerminalService';
+import { WindowsExternalTerminalService, MacExternalTerminalService, LinuxExternalTerminalService } from 'vs/platform/externalTerminal/node/externalTerminalService';
 
 suite('ExternalTerminalService', () => {
 	let mockOnExit: Function;
