+++ src/vs/platform/externalTerminal/node/externalTerminalService.ts
