+++ src/vs/workbench/contrib/externalTerminal/node/terminals.ts
@@ -6,6 +6,10 @@
 import * as cp from 'child_process';
 import * as platform from 'vs/base/common/platform';
 import { getDriveLetter } from 'vs/base/common/extpath';
+import { LinuxExternalTerminalService, MacExternalTerminalService, WindowsExternalTerminalService } from 'vs/platform/externalTerminal/node/externalTerminalService';
+import { IExternalTerminalService } from 'vs/platform/externalTerminal/common/externalTerminal';
+import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
+import { ExtHostConfigProvider } from 'vs/workbench/api/common/extHostConfiguration';
 
 
 function spawnAsPromised(command: string, args: string[]): Promise<string> {
@@ -24,6 +28,24 @@ function spawnAsPromised(command: string, args: string[]): Promise<string> {
 			resolve(stdout);
 		});
 	});
+}
+
+let externalTerminalService: IExternalTerminalService | undefined = undefined;
+
+export function runInExternalTerminal(args: DebugProtocol.RunInTerminalRequestArguments, configProvider: ExtHostConfigProvider): Promise<number | undefined> {
+	if (!externalTerminalService) {
+		if (platform.isWindows) {
+			externalTerminalService = new WindowsExternalTerminalService(<IConfigurationService><unknown>undefined);
+		} else if (platform.isMacintosh) {
+			externalTerminalService = new MacExternalTerminalService(<IConfigurationService><unknown>undefined);
+		} else if (platform.isLinux) {
+			externalTerminalService = new LinuxExternalTerminalService(<IConfigurationService><unknown>undefined);
+		} else {
+			throw new Error('external terminals not supported on this platform');
+		}
+	}
+	const config = configProvider.getConfiguration('terminal');
+	return externalTerminalService.runInTerminal(args.title!, args.cwd, args.args, args.env || {}, config.external || {});
 }
 
 export function hasChildProcesses(processId: number | undefined): Promise<boolean> {
